% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme_from_sntask.R
\name{create_readme_from_sntask}
\alias{create_readme_from_sntask}
\title{Create a README from a task}
\usage{
create_readme_from_sntask(task, path)
}
\arguments{
\item{task}{an object of sntask}

\item{path}{Path or connection to write to}
}
\value{
invisible task
}
\description{
Create a README from a task
}
