% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sntask.R
\name{read_sntask}
\alias{read_sntask}
\title{Read a Task PDF from ServiceNow}
\usage{
read_sntask(file)
}
\arguments{
\item{file}{Path to PDF}
}
\value{
a sntask() object
}
\description{
Read a Task PDF from ServiceNow
}
